+++ provider/database_autonomous_database_test.go
@@ -121,7 +121,6 @@ func TestDatabaseAutonomousDatabaseResource_basic(t *testing.T) {
 				),
 			},
 			// verify updates to updatable parameters
-			// @CODEGEN 09/2018 - Updating password and other fields except cpu_core_count and data_storage_size_in_tbs as service does not support scaling and password update at same time
 			{
 				Config: config + `
 variable "autonomous_database_admin_password" { default = "BEstrO0ng_#12" }
@@ -157,50 +156,12 @@ variable "autonomous_database_state" { default = "AVAILABLE" }
 					},
 				),
 			},
-			// verify updates to updatable parameters - scaling properties: cpu_core_count and data_storage_size_in_tbs
-			// @CODEGEN 09/2018 - Despite scaling the ADB successfully, cpu_core_count and data_storage_size_in_tbs values returned may still return original values resulting in plan diff
-			{
-				Config: config + `
-variable "autonomous_database_admin_password" { default = "BEstrO0ng_#12" }
-variable "autonomous_database_cpu_core_count" { default = 2 }
-variable "autonomous_database_data_storage_size_in_tbs" { default = 2 }
-variable "autonomous_database_db_name" { default = "adatabasedb1" }
-variable "autonomous_database_defined_tags_value" { default = "updatedValue" }
-variable "autonomous_database_display_name" { default = "displayName2" }
-variable "autonomous_database_freeform_tags" { default = {"Department"= "Accounting"} }
-variable "autonomous_database_license_model" { default = "LICENSE_INCLUDED" }
-variable "autonomous_database_state" { default = "AVAILABLE" }
-
-                ` + compartmentIdVariableStr + AutonomousDatabaseResourceConfig,
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceName, "admin_password", "BEstrO0ng_#12"),
-					resource.TestCheckResourceAttr(resourceName, "compartment_id", compartmentId),
-					resource.TestCheckResourceAttrSet(resourceName, "cpu_core_count"),
-					resource.TestCheckResourceAttrSet(resourceName, "data_storage_size_in_tbs"),
-					resource.TestCheckResourceAttr(resourceName, "db_name", "adatabasedb1"),
-					resource.TestCheckResourceAttr(resourceName, "defined_tags.%", "1"),
-					resource.TestCheckResourceAttr(resourceName, "display_name", "displayName2"),
-					resource.TestCheckResourceAttr(resourceName, "freeform_tags.%", "1"),
-					resource.TestCheckResourceAttrSet(resourceName, "id"),
-					resource.TestCheckResourceAttr(resourceName, "license_model", "LICENSE_INCLUDED"),
-					resource.TestCheckResourceAttrSet(resourceName, "state"),
-
-					func(s *terraform.State) (err error) {
-						resId2, err = fromInstanceState(s, resourceName, "id")
-						if resId != resId2 {
-							return fmt.Errorf("Resource recreated when it was supposed to be updated.")
-						}
-						return err
-					},
-				),
-				ExpectNonEmptyPlan: true,
-			},
 			// verify datasource
 			{
 				Config: config + `
 variable "autonomous_database_admin_password" { default = "BEstrO0ng_#12" }
-variable "autonomous_database_cpu_core_count" { default = 2 }
-variable "autonomous_database_data_storage_size_in_tbs" { default = 2 }
+variable "autonomous_database_cpu_core_count" { default = 1 }
+variable "autonomous_database_data_storage_size_in_tbs" { default = 1 }
 variable "autonomous_database_db_name" { default = "adatabasedb1" }
 variable "autonomous_database_defined_tags_value" { default = "updatedValue" }
 variable "autonomous_database_display_name" { default = "displayName2" }
@@ -229,8 +190,8 @@ data "oci_database_autonomous_databases" "test_autonomous_databases" {
 
 					resource.TestCheckResourceAttr(datasourceName, "autonomous_databases.#", "1"),
 					resource.TestCheckResourceAttr(datasourceName, "autonomous_databases.0.compartment_id", compartmentId),
-					resource.TestCheckResourceAttrSet(datasourceName, "autonomous_databases.0.cpu_core_count"),
-					resource.TestCheckResourceAttrSet(datasourceName, "autonomous_databases.0.data_storage_size_in_tbs"),
+					resource.TestCheckResourceAttr(datasourceName, "autonomous_databases.0.cpu_core_count", "1"),
+					resource.TestCheckResourceAttr(datasourceName, "autonomous_databases.0.data_storage_size_in_tbs", "1"),
 					resource.TestCheckResourceAttr(datasourceName, "autonomous_databases.0.db_name", "adatabasedb1"),
 					resource.TestCheckResourceAttr(datasourceName, "autonomous_databases.0.defined_tags.%", "1"),
 					resource.TestCheckResourceAttr(datasourceName, "autonomous_databases.0.display_name", "displayName2"),
@@ -239,14 +200,13 @@ data "oci_database_autonomous_databases" "test_autonomous_databases" {
 					resource.TestCheckResourceAttr(datasourceName, "autonomous_databases.0.license_model", "LICENSE_INCLUDED"),
 					resource.TestCheckResourceAttrSet(datasourceName, "autonomous_databases.0.state"),
 				),
-				ExpectNonEmptyPlan: true,
 			},
 			// verify singular datasource
 			{
 				Config: config + `
 variable "autonomous_database_admin_password" { default = "BEstrO0ng_#12" }
-variable "autonomous_database_cpu_core_count" { default = 2 }
-variable "autonomous_database_data_storage_size_in_tbs" { default = 2 }
+variable "autonomous_database_cpu_core_count" { default = 1 }
+variable "autonomous_database_data_storage_size_in_tbs" { default = 1 }
 variable "autonomous_database_db_name" { default = "adatabasedb1" }
 variable "autonomous_database_defined_tags_value" { default = "updatedValue" }
 variable "autonomous_database_display_name" { default = "displayName2" }
@@ -264,8 +224,8 @@ data "oci_database_autonomous_database" "test_autonomous_database" {
 
 					resource.TestCheckResourceAttr(singularDatasourceName, "compartment_id", compartmentId),
 					resource.TestCheckResourceAttrSet(singularDatasourceName, "connection_strings.#"),
-					resource.TestCheckResourceAttrSet(singularDatasourceName, "cpu_core_count"),
-					resource.TestCheckResourceAttrSet(singularDatasourceName, "data_storage_size_in_tbs"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "cpu_core_count", "1"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "data_storage_size_in_tbs", "1"),
 					resource.TestCheckResourceAttr(singularDatasourceName, "db_name", "adatabasedb1"),
 					resource.TestCheckResourceAttr(singularDatasourceName, "defined_tags.%", "1"),
 					resource.TestCheckResourceAttr(singularDatasourceName, "display_name", "displayName2"),
@@ -275,7 +235,6 @@ data "oci_database_autonomous_database" "test_autonomous_database" {
 					resource.TestCheckResourceAttr(singularDatasourceName, "state", "AVAILABLE"),
 					resource.TestCheckResourceAttrSet(singularDatasourceName, "time_created"),
 				),
-				ExpectNonEmptyPlan: true,
 			},
 			// verify resource import
 			{